DGGS._transform_points — Method

Transforms Vector of (lon,lat) coords to DGGRID indices

DGGS.aggregate_dggs_layer — Method

Spatial hexagonal convolution in Q2DI index space matching levels of DGGRID ISEA4H grids

DGGS.call_dggrid — Method

Execute sytem call of DGGRID binary

DGGS.filter_null — Method

Apply function f after filtering of missing and NAN values

DGGS.to_dggs_layer — Method

Transforms a YAXArrays.Dataset in geographic lat/lon ratser to a DGGSLayer at agiven layer

DGGS.transform_points — Method

chunksizepoints: number of points (e.g. pixels) to transform in one block (task of a thread)

DGGS.width — Method

position of last row or column in a quad matrix of that level