DGGS.get_cell_cube — Function

Import geographical data cube into a DGGS

Transforms a data cube with spatial index dimensions longitude and latitude into a data cube with the cell id as a single spatial index dimension. Re-gridding is done using the average value of all geographical coordinates belonging to a particular cell defined by the grid specification grid_spec.

DGGS.get_cell_ids — Method

Convert geographic corrdinates to cell id

DGGS.get_geo_coords — Method

Convert cell id to geographic coordinate of cell center

DGGS.get_geo_cube — Method

Export cell data cube into a traditional geographical one

Transforms a data cube with one spatial index dimensions, i. e., the cell id, into a traditional geographical data cube with two spatial index dimensions longitude and latitude. Re-gridding is done by putting the cell value into the center point of its cell.