DGGS.PresetGridSpecs — Constant

DGGRID grid presets

DGGS.Grid — Type

A grid as defined by DGGRID with the actual grid points

DGGS.Grid — Method

Create a grid using DGGRID parameters

DGGS.Grid — Method

Create a grid using DGGRID preset

DGGS.GridSpec — Type

Specification of a grid as defined by DGGRID

DGGS.call_dggrid — Method

Execute sytem call of DGGRID binary

DGGS.create_grids — Method

Create a grid system of different resolutions

DGGS.get_cell_boundaries — Method

Get a GeoDataFrame with boundary polygons for each cell

DGGS.get_cell_cube — Method

Import geographical data cube into a DGGS

Transforms a data cube with spatial index dimensions longitude and latitude into a data cube with the cell id as a single spatial index dimension. Re-gridding is done using the average value of all geographical coordinates belonging to a particular cell defined by the grid specification grid_spec.

DGGS.get_cell_ids — Method

Convert geographic corrdinates to cell id

DGGS.get_cell_ids — Method

Transform geographical coordinates into cell ids

DGGS.get_children_cell_ids — Method

Get ids of all child cells

DGGS.get_cube_pyramid — Method

Get a cell data cube pyramid

Calculates a stack of cell data cubes with incrementally lower resolutions based on the same data as provided by cell_cube

DGGS.get_geo_coords — Method

Convert cell id to geographic coordinate of cell center

DGGS.get_geo_cube — Method

Export cell data cube into a traditional geographical one

Transforms a data cube with one spatial index dimensions, i. e., the cell id, into a traditional geographical data cube with two spatial index dimensions longitude and latitude. Re-gridding is done by mapping the cell value into the center point of its cell. Interpolation of values is performed, because grid cells on the same meridian may have different latitudes.

DGGS.get_grid_data — Method

Get a DataFrame of cell center points

DGGS.get_kd_tree — Method

represent cell table as kd-tree of center points

DGGS.get_parent_cell_id — Method

Get id of parent cell