DGGS.Apertures — Constant

DGGRID apertures

DGGS.Projections — Constant

DGGRID projections

DGGS.Topologies — Constant

DGGRID topologies

DGGS.AbstractCube — Type

Subtypes must have field data::YAXArray

DGGS.AbstractGrid — Type

A set of points defining a grid as a voronoi partition. Points must be geographical coordinates (lon, lat) and stored as a KDTree in the filed data.

DGGS.CellCube — Method

Import geographical data cube into a DGGS

Transforms a data cube with spatial index dimensions longitude and latitude into a data cube with the cell id as a single spatial index dimension. Re-gridding is done using the average value of all geographical coordinates belonging to a particular cell defined by the grid specification grid_spec.

DGGS.DgGrid — Method

Create a grid using DGGRID parameters

DGGS.GeoCube — Method

Export cell data cube into a traditional geographical one

Transforms a data cube with one spatial index dimensions, i. e., the cell id, into a traditional geographical data cube with two spatial index dimensions longitude and latitude. Values are taken from the nearest cell.

DGGS.Grid — Method

Create a Grid using coordinates of center points describing a voronoi partition. center_points must have one point per column with 2 rows for longitude and latitude, respectiveley.

DGGS.call_dggrid — Method

Execute sytem call of DGGRID binary

DGGS.get_cell_boundaries — Method

Get a GeoDataFrame with boundary polygons for each cell

DGGS.get_cell_ids — Method

Get cell ids given geographic corrdinates

DGGS.get_children_cell_ids — Method

Get the cell ids of the children associated to each cell id of the parent level

Note that a child often has multiple parents. The cells at different resolutions are only in rectangular pyramids perfectly nested. Returns a Dict with parent cell id as keys and children cell ids as values

DGGS.get_cube_pyramid — Method

Get a cell data cube pyramid

Calculates a stack of cell data cubes with incrementally lower levels based on the same data as provided by cell_cube. Cell values are combined according to the provided aggregate_function.

DGGS.get_geo_coords — Method

Convert cell id to geographic coordinate of cell center

DGGS.get_geo_coords — Method

Convert cell id to geographic coordinate of cell center

DGGS.get_grid_table — Method

Get a DataFrame of cell center points

DGGS.knn — Method

Get cell ids of k nearest neighbors arround a cell